feat: Update for ColdBox 5 and cbstorages 2.0.0
New settings introduced: `cookieCheck` which defaults to `true` to check
for a CFID cookie to determine if the request is likely a bot and
`botsJsonFile` to change what file to load.  Additionally, the configured
`botsJsonFile` is loaded and mapped to `Bots@disable-bot-sessions` in the
`ModuleConfig.cfc`.  Also added cfformat support and expanded CI matrix testing.

BREAKING CHANGE:  Support dropped for ACF 11 and Lucee 4.5